Refactor: Isolate the shared browser_engine class code
